Experiment with Jinja2
(https://github.com/mitsuhiko/jinja2)

The good parts:
 - as expected, it's **fast**, between 5x to 9x faster than Genshi,
   on par with Kajiki
 - the syntax is not that close from Genshi, however:
   - it's customizable, so we can get back the Genshi expressions
     `${ ... }`
   - it has getattr/getitem unification, so no ugly `dict['key']`
     everywhere needed, like with Kajiki
   - it's quite straightforward to adapt the <py:choose> with
     if/elif/else/endif
   - nice default handling of undefined values, just what we need
   - we heavily use the `Line Statements`, even for HTML.
     This makes it very easy to visually separate the template content
     from the template logic, as the later stands out on its own lines,
     in a distinctive style. It's actually much clearer than with
     Genshi / Kajiki where everything is XML, and I believe it's also
     clearer than the traditional block syntax
   - when we need to use the block syntax, we use `{{ ... }}`
     instead of the ugly `{% ... %}`, but this usage is reserved
     for one-liner control structures
     (e.g. `{{ if fullrow }}<hr />{{ endif }}`)
 - everything works as documented and as expected (except for one glitch
   with block assignments, issue 490)
 - the error reporting is **very** effective
 - the `extends` works with a variable value, so we can do themes
 - the `include` supports missing templates, and can takes variables
   from the context, so <xi:include>s are easy to adapt
 - Jinja2 doesn't know anything about HTML/XML:
   - so it won't stop you from having template control structures inside
     <script> elements...  (can't do that with Kajiki) -
   - but it nevertheless has some useful filters, like `xmlattr`

The bad parts:
 - quite some work ahead, but we knew that before starting...

Jinja2 vs. Kajiki: despite the initial appeal of Kajiki with its
syntax close to the Genshi one, there's still a lot of adaptations to
do, some much more cumbersome to do in Kajiki than in Jinja2
(<py:choose>, parametric <xi:include>). Also, mostly everything I
tried in Jinja2 worked as expected, not so with Kajiki. And when
something goes wrong, the error messages and stack traces in Jinja2
were **always** helpful and to the point, so Jinja2 templates are much
faster to develop and maintain as a result.

Both are in the same league speed wise, and Genshi is in the dust ;-)

Didn't benchmark yet the memory footprint, nor the streaming mode.

Still TODO: fix/check the i18n, define alternatives for the Genshi
stream filters we use.