jinja2: rework template blocks
The default root of the template hierarchy is the theme template
(`jtheme.html` by default).  It contains the HTML structure, the
<head> and <body> blocks.  These HTML tags contain also corresponding
Jinja2 blocks, "head" and "body". In addition, the body block contains
a "bottom" block, which comes after the whole standard navigation
controls.

The second level of the template hierarchy is layout, which extends
the current theme. It fills the "head" block with the standard set of
<meta>, <script> and other head tags seen on all pages, and fills the
"bottom" block with a "content" block, followed by the late scripts,
and the links for downlloads in alternate formats.

The third level is typically the module specific page, which extends
the layout and provides its own supplemental content for the head
block (calling `${super()}` first), and its own content for the
"content" block.

There can be additional levels, like for the admin (or preference)
panels which will likely extend an admin (pref) page.

Additional changes:
 - add copyright notice (Kajiki had trouble with it)
 - in jlayout.html, add late links and late scripts that were somehow
   missed in the initial migration
 - in jtheme.html, same thing happened to the footer