theme and layout: simplify the block structure
Upon reflexion, we don't loose in flexibility when directly using the
`content` block in the jtheme.html page (getting rid of the
intermediate `bottom` level).

The only additional constraint is to have to remember to include a
call to `${ super() }` in the end user templates.  However, this is
both more explicit and more flexible, as it could also be omitted on
purpose.