query: port the Genshi template to Jinja2
Including batch modification, and even the !TicketQuery macro works
now with Jinja2. The `late_scripts` stuff seems to work as well
(tested with a ticket containing `[[Workflow]]` in its description,
then showing it in the results).

(the above should have been the commit message for 1678126f)

Also added jquery.rss, the migration of query.rss.