ticket box: correctly scope `fields`
Before, without the `with` directive, the modification
to the `fields` would ''spread'' to the other use of
`fields`, for example in the `fields[fields_map[...]]`
expression in the <h2>, where it will cause errors.
It's a bit surprising that it does so, but probably the
set expressions are evaluated at the beginning of their
scope, which would explain the behavior.