wiki: fix the last unit tests.
These last failures were tricky ones, as testing via `make test-wiki`
would work, it's only when running `make unit-test` that the
''!WikiProcessor code block with invalid arguments'' test and related
ones would fail.

This was due to the `PygmentsRenderer` being used for rendering
`text/plain` mimetype in the failing case, instead of the
`PlainTextRenderer` in the succeeding case.

As seen before, Pygments escapes quotes (") as `&quot;`.

The HTML escaping in the case of the `PlainTextRenderer` is done by
the tag builder, which uses `escape(..., False)`, i.e. don't escape
the quotes. If we would use `escape(..., True)` the quotes would be
escaped as `&34;` and would make numerous other test fail. So we
preferred to force using `PlainTextRenderer` for these tests, even
when `PygmentsRenderer` is present.