html: serialization of Fragments wasn't correct
We could generate buggy HTML, as seen in the ticket admin tests with
context navigation content like:
{{{
<a class="next" href="/trunk-tracenv/milestone/milestone2" title="Milestone "milestone2"">
}}}
This was detected by the XML validation done by lxml.

With `escape` passing through its Fragment input serialized but
otherwise unmodified, we get now the expected:
{{{
<a class="next" href="/trunk-tracenv/milestone/milestone2" title="Milestone &#34;milestone2&#34;">
}}}
(previously, the `unicode(e).replace('&#34;', '"')` was
undiscriminately modifying the content of the attributes).

We also remove the redundant `Element.__html__` method, as it's
currently the same as the `Fragment.__html__` one.